Adds dev config overrides
